Move files back to their original folders
