+++ autogpts/autogpt/autogpt/agents/prompt_strategies/one_shot.py
