+++ autogpts/autogpt/autogpt/commands/git_operations.py
@@ -5,7 +5,7 @@ from git.repo import Repo
 
 from autogpt.agents.protocols import CommandProvider
 from autogpt.utils.exceptions import CommandExecutionError
-from autogpt.utils.command_decorator import command
+from autogpt.command_decorator import command
 from autogpt.config.config import Config
 from autogpt.core.utils.json_schema import JSONSchema
 from autogpt.models.command import Command
