+++ autogpts/autogpt/autogpt/commands/image_gen.py
@@ -14,7 +14,7 @@ from openai import OpenAI
 from PIL import Image
 
 from autogpt.agents.protocols import CommandProvider
-from autogpt.utils.command_decorator import command
+from autogpt.command_decorator import command
 from autogpt.config.config import Config
 from autogpt.core.utils.json_schema import JSONSchema
 from autogpt.file_storage.base import FileStorage
