+++ autogpts/autogpt/autogpt/commands/web_search.py
@@ -7,7 +7,7 @@ from duckduckgo_search import DDGS
 
 from autogpt.agents.protocols import CommandProvider, DirectiveProvider
 from autogpt.utils.exceptions import ConfigurationError
-from autogpt.utils.command_decorator import command
+from autogpt.command_decorator import command
 from autogpt.config.config import Config
 from autogpt.core.utils.json_schema import JSONSchema
 from autogpt.models.command import Command
